+++ lib/views/main_view/main_view_widgets/category_widgets/token_category_actions.dart/delete_token_category_action.dart
@@ -30,6 +30,7 @@ class DeleteTokenCategoryAction extends StatelessWidget {
       context: globalNavigatorKey.currentContext!,
       builder: (BuildContext context) {
         return AlertDialog(
+          scrollable: true,
           title: Text(
             AppLocalizations.of(context)!.confirmDeletion,
           ),
