+++ lib/views/main_view/main_view_widgets/category_widgets/token_category_actions.dart/rename_token_category_action.dart
@@ -32,6 +32,7 @@ class RenameTokenCategoryAction extends StatelessWidget {
         context: globalNavigatorKey.currentContext!,
         builder: (BuildContext context) {
           return AlertDialog(
+            scrollable: true,
             title: Text(
               AppLocalizations.of(context)!.renameTokenCategory,
             ),
