+++ lib/views/main_view/main_view_widgets/token_widgets/day_password_token_widgets/day_password_token_widget.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 
 import '../../../../../model/tokens/day_password_token.dart';
 import '../token_widget.dart';
-import 'actions/edit_day_password_token_action.dart';
 import '../token_widget_base.dart';
+import 'actions/edit_day_password_token_action.dart';
 import 'day_password_token_widget_tile.dart';
 
 class DayPasswordTokenWidget extends TokenWidget {
