+++ lib/views/main_view/main_view_widgets/token_widgets/default_token_actions/default_delete_action.dart
@@ -34,6 +34,7 @@ class DefaultDeleteAction extends TokenAction {
       context: globalNavigatorKey.currentContext!,
       builder: (BuildContext context) {
         return AlertDialog(
+          scrollable: true,
           title: Text(
             AppLocalizations.of(context)!.confirmDeletion,
           ),
