+++ lib/views/main_view/main_view_widgets/token_widgets/default_token_actions/default_edit_action.dart
@@ -35,6 +35,7 @@ class DefaultEditAction extends TokenAction {
         context: globalNavigatorKey.currentContext!,
         builder: (BuildContext context) {
           return AlertDialog(
+            scrollable: true,
             title: Text(
               AppLocalizations.of(context)!.renameToken,
               overflow: TextOverflow.fade,
