+++ lib/views/main_view/main_view_widgets/token_widgets/push_token_widgets/rollout_failed_widget.dart
@@ -66,6 +66,7 @@ class RolloutFailedWidget extends StatelessWidget {
       context: globalNavigatorKey.currentContext!,
       builder: (BuildContext context) {
         return AlertDialog(
+          scrollable: true,
           title: Text(
             AppLocalizations.of(context)!.confirmDeletion,
           ),
