+++ lib/views/main_view/main_view_widgets/token_widgets/token_widget_base.dart
@@ -2,9 +2,8 @@ import 'dart:math';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import '../../../../model/mixins/sortable_mixin.dart';
-import 'token_widget_slideable.dart';
 
+import '../../../../model/mixins/sortable_mixin.dart';
 import '../../../../model/tokens/token.dart';
 import '../../../../utils/riverpod_providers.dart';
 import '../../../../utils/text_size.dart';
@@ -12,6 +11,7 @@ import 'default_token_actions/default_delete_action.dart';
 import 'default_token_actions/default_edit_action.dart';
 import 'default_token_actions/default_lock_action.dart';
 import 'token_action.dart';
+import 'token_widget_slideable.dart';
 
 class TokenWidgetBase extends ConsumerWidget {
   final Widget tile;
