+++ lib/views/main_view/main_view_widgets/token_widgets/totp_token_widgets/totp_token_widget.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import '../../../../../model/mixins/sortable_mixin.dart';
 import '../../../../../model/tokens/totp_token.dart';
 import '../token_widget.dart';
-import 'actions/edit_totp_token_action.dart';
 import '../token_widget_base.dart';
+import 'actions/edit_totp_token_action.dart';
 import 'totp_token_widget_tile.dart';
 
 class TOTPTokenWidget extends TokenWidget {
